+++ diracx-core/src/diracx/core/config/sources.py
@@ -9,11 +9,10 @@ import asyncio
 import logging
 import os
 from abc import ABCMeta, abstractmethod
-from dataclasses import dataclass
 from datetime import datetime, timezone
 from pathlib import Path
 from tempfile import TemporaryDirectory
-from typing import Annotated, ClassVar, Generic, TypeVar
+from typing import Annotated, Generic, TypeVar
 from urllib.parse import urlparse, urlunparse
 
 import sh
@@ -23,7 +22,7 @@ from pydantic import AnyUrl, BeforeValidator, TypeAdapter, UrlConstraints
 
 from diracx.core.exceptions import BadConfigurationVersionError
 from diracx.core.extensions import DiracEntryPoint, select_from_extension
-from diracx.core.utils import AsyncTwoLevelCache, TwoLevelCache
+from diracx.core.utils import TwoLevelCache
 
 from .schema import Config
 
@@ -138,79 +137,6 @@ class CacheableSource(Generic[T], metaclass=ABCMeta):
         """Clear the caches."""
         self._revision_cache.clear()
         self._content_cache.clear()
-
-
-@dataclass(frozen=True)
-class Snapshot(Generic[T]):
-    """Wraps a cached data payload with its cache metadata."""
-
-    data: T
-    hexsha: str
-    modified: datetime
-
-
-class AsyncCacheableSource(Generic[T], metaclass=ABCMeta):
-    """Abstract base class for async sources that can be cached.
-
-    Async equivalent of CacheableSource. Uses AsyncTwoLevelCache so populate
-    functions are native coroutines.
-    """
-
-    #: The database class this source reads from. Used by the application
-    #: factory to instantiate the source with the matching database instance.
-    db_class: ClassVar[type]
-
-    def __init__(self):
-        self._revision_cache = AsyncTwoLevelCache(
-            soft_ttl=DEFAULT_CS_REV_CACHE_SOFT_TTL,
-            hard_ttl=DEFAULT_CS_REV_CACHE_HARD_TTL,
-            max_items=1,
-        )
-        self._content_cache: Cache = LRUCache(maxsize=2)
-
-    @abstractmethod
-    async def latest_revision(self) -> tuple[str, datetime]:
-        """Return (revision_str, modified) identifying the current revision."""
-
-    @abstractmethod
-    async def read_raw(self, hexsha: str, modified: datetime) -> T:
-        """Fetch and return the data for the given revision."""
-
-    async def _read_work(self) -> str:
-        hexsha, modified = await self.latest_revision()
-        if hexsha not in self._content_cache:
-            self._content_cache[hexsha] = await self.read_raw(hexsha, modified)
-        return hexsha
-
-    async def read(self) -> T:
-        """Blocking read — awaits refresh on a hard cache miss."""
-        hexsha = await self._revision_cache.get(
-            "latest_revision", self._read_work, blocking=True
-        )
-        return self._content_cache[hexsha]
-
-    async def read_non_blocking(self) -> T:
-        """Non-blocking read — raises NotReadyError on a hard cache miss."""
-        hexsha = await self._revision_cache.get(
-            "latest_revision", self._read_work, blocking=False
-        )
-        return self._content_cache[hexsha]
-
-    async def clear_caches(self):
-        """Clear the caches."""
-        await self._revision_cache.clear()
-        self._content_cache.clear()
-
-    @classmethod
-    async def create(cls) -> T:
-        """Dependency injection stub.
-
-        The application factory instantiates each concrete source and
-        overrides ``cls.create`` with the instance's ``read`` method, so this
-        should never actually be called. Each subclass's bound ``create``
-        classmethod is a distinct dependency key.
-        """
-        raise NotImplementedError(f"{cls.__name__} was not wired by the factory")
 
 
 class ConfigSource(CacheableSource[Config]):
